Removes most of special char handling.
The special char handling was introduced in order to have a stronger
prefix base. This is handling is now removed because the output texts
after this handling do not look any longer like tweets.

Another solution is needed but not yet at hand

URL-encoded special chars like < or & are still handled