Introduces Prefix object, seedable Random and adjustable window size
A larger part of the TweetkovChain class deals with prefix handling.
This commit moves operations to a new Prefix class where it belongs in
terms of business logic.

This move is supported by a diversity of tests which look at the two
major window sizes (aka Markov Orders), 1 and 2.

Furthermore the original case of the tokens is kept in order to have it
look more natural.

The parts that use the pseudo-random generator are now seeded in order
to test the sentence generation.