Feature #2: Adds support to remove duplicate sentences
Duplicate sentences arise from data where each link (aka Transition)
from prefix to suffix in the whole chain has the selection probability
of p=1.0 (all suffixes in that sentence are always selected).

This commit adds support to remove such sentences (aka "duplicates")
on a best effort basis. The idea is, to track transitions with a suffix
probability of p=1.0. During the sentence generation such a sentence can
be detected. Then a non-duplicate replace can be generated (aka
"alternative")

A retry mechanism takes care of avoiding sentences (in terms of the
contract of avoiding duplicates) in order to generate a sentence. If
during a retry a defined retry-threshold is hit then an warning is
logged and the method returns with an empty sentence.

There is still a high change that a duplicate can be generated. This is
because the original tweets are lost during the tokenization. The only
way to definitely know if a duplicate is at hand is the number of unique
transitions (as mentioned above). Once there is the sheer possibility to
select an alternative transition the whole sentence is viewed as
alternative.

Please note that the older method of generation (with duplicates) is
still active and running.

This commit moves the training data into its own repository class in
order to have a point of segregation so the actual implementation may
be exchanged. Transition is now the class which maps a prefix to any
number of suffixes.

Furthermore, a lot of logic moves from TweetkovChain into new classes
where the handling should belong. This also means that initializeSeed()
trails now through the chain of command in order to enable testing with
pseudo-random generators.

printHistogramm is converted to generateHistogram so users can decide
what to do with the histogram. It is still a single string though.